Fix deployment and add Express interfaces
