+++ src/metrics/typings.json
@@ -4,6 +4,10 @@
     "csv-parse": "registry:dt/csv-parse#1.1.0+20160602150508",
     "es6-collections": "registry:dt/es6-collections#0.5.1+20160316155526",
     "es6-promise": "registry:dt/es6-promise#0.0.0+20160614011821",
-    "node-4": "registry:dt/node-4#4.0.0+20160801161248"
+    "express": "registry:dt/express#4.0.0+20160708185218",
+    "express-serve-static-core": "registry:dt/express-serve-static-core#4.0.0+20160715232503",
+    "mime": "registry:dt/mime#0.0.0+20160316155526",
+    "node-4": "registry:dt/node-4#4.0.0+20160801161248",
+    "serve-static": "registry:dt/serve-static#0.0.0+20160606155157"
   }
 }
